% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycrime.R
\name{my_api_key}
\alias{my_api_key}
\title{NYC CRIME DATA}
\usage{
my_api_key(your_API = "FJPsznoKWRmpTORBUtMkarl17")
}
\description{
The functions are designed to let users
get access to the NYC arrest data from
01, 2019 to 09, 2019.
}
\details{
Also, the functions can help the users
generate a clearer and easier way to
get to know more information for some variables.
}
